# Switch to mesh rasterization and screen-space reflections