--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 maps/
+out/
 bin/
 *.swo
 *.swp