--- a/src/map.js
+++ b/src/map.js
@@ -168,10 +168,8 @@ async function main() {
     )
 
     gl.enable(gl.DEPTH_TEST)
-    /*
     gl.enable(gl.CULL_FACE)
     gl.cullFace(gl.BACK)
-    */
 
     gl.useProgram(program)